test(queries): resettled test_select_renderer to separate test files
